leverage new method to hide data in other logs
