+++ unstructured/ingest/cli/common.py
@@ -1,10 +1,10 @@
-import json
 import logging
 from typing import Optional
 
 from click import ClickException
 
 from unstructured.ingest.logger import ingest_log_streaming_init, logger
+from unstructured.ingest.utils.logging import hide_sensitive_fields
 
 
 def run_init_checks(
@@ -57,33 +57,6 @@ def run_init_checks(
         logger.warning(
             "Not preserving downloaded files but download_dir is specified",
         )
-
-
-def hide_sensitive_fields(data: dict) -> dict:
-    sensitive_fields = [
-        "account_name",
-        "client_id",
-    ]
-    sensitive_triggers = ["key", "cred", "token"]
-    new_data = data.copy()
-    for k, v in new_data.items():
-        if (
-            any([s in k.lower() for s in sensitive_triggers])  # noqa: C419
-            or k.lower() in sensitive_fields
-        ):
-            new_data[k] = "*******"
-        if isinstance(v, dict):
-            new_data[k] = hide_sensitive_fields(v)
-        if isinstance(v, str):
-            try:
-                json_data = json.loads(v)
-                if isinstance(json_data, dict):
-                    updated_data = hide_sensitive_fields(json_data)
-                    new_data[k] = json.dumps(updated_data)
-            except json.JSONDecodeError:
-                pass
-
-    return new_data
 
 
 def log_options(options: dict, verbose=False):
